optimize
Before querying, determine the number of candidate vectors based on the index.
In the process of calculating the nearest neighbor, the number of vectors in the cluster determines whether to use brute force search or search with an "early termination" algorithm.
Fixed the issue of low hit rate caused by always deleting the last result when summarizing nearest neighbors.