optimize
Optimized the algorithm for querying the nearest neighbors of each layer.
Reconstructed the structure of the project.
Fixed the issue of not clearing the "recent distance" when calculating the nearest neighbor if the distance between the vector in the waiting queue and the query vector is less than the maximum value in the priority queue.