Changed the logic of all search related methods.
In stage one, when quickly approaching the target vertex through a long edge, it is not enough to use only the long edge.

Let's set the current vertex, short edge neighbor, target vertex, and long edge neighbor to four points and one line, sorted from left to right. The distance between the short edge neighbor and the current vertex is 100, the distance between the long edge neighbor and the current vertex is 150, and the target vertex is 120 (20 from the short edge neighbor and 30 from the long edge neighbor). In this case, the distance between the target vertex and the current vertex cannot represent the distance between the target vertex and its short edge neighbors. If only long edge neighbors are used, detours may occur.